+++ benchmarks/internal/cluster/dynamic.js
@@ -17,7 +17,7 @@ const dynamicPoolLessRecentlyUsed = new DynamicClusterPool(
   size / 2,
   size * 3,
   './benchmarks/internal/cluster/worker.js',
-  { workerChoiceStrategy: WorkerChoiceStrategies.LESS_RECENTLY_USED }
+  { workerChoiceStrategy: WorkerChoiceStrategies.LESS_USED }
 )
 
 const dynamicPoolWeightedRoundRobin = new DynamicClusterPool(