+++ benchmarks/internal/thread/fixed.js
@@ -15,7 +15,7 @@ const fixedPool = new FixedThreadPool(
 const fixedPoolLessRecentlyUsed = new FixedThreadPool(
   size,
   './benchmarks/internal/thread/worker.js',
-  { workerChoiceStrategy: WorkerChoiceStrategies.LESS_RECENTLY_USED }
+  { workerChoiceStrategy: WorkerChoiceStrategies.LESS_USED }
 )
 
 const fixedPoolWeightedRoundRobin = new FixedThreadPool(