+++ src/pools/selection-strategies/less-used-worker-choice-strategy.ts
@@ -2,13 +2,13 @@ import type { IPoolWorker } from '../pool-worker'
 import { AbstractWorkerChoiceStrategy } from './abstract-worker-choice-strategy'
 
 /**
- * Selects the less recently used worker.
+ * Selects the less used worker.
  *
  * @typeParam Worker - Type of worker which manages the strategy.
  * @typeParam Data - Type of data sent to the worker. This can only be serializable data.
  * @typeParam Response - Type of response of execution. This can only be serializable data.
  */
-export class LessRecentlyUsedWorkerChoiceStrategy<
+export class LessUsedWorkerChoiceStrategy<
   Worker extends IPoolWorker,
   Data,
   Response